Made link to issues absolute
Link to issues is currently broken. This fixes it by making it absolute.

As referenced here https://help.github.com/articles/about-readmes/ relative links are parsed to be relative to the root of the repo so people can link to specific files.

I think its probably reliable enough to use the absolute path.