Update exif.js: Allow image from uploaded blobs
Allow not completed for the reason, that If one generates images from blob and creates image element from it: EXIF's datas are there BUT the image element does not show up as complete.

Nice and non-noisy fix is that the developer has opportunity to 'fake' complete flag setting. It seems to work then ok, no extra change needs.